IGNITE-23126 Re-factor tests
