+++ modules/platforms/python/tests/conftest.py
@@ -14,12 +14,32 @@
 # limitations under the License.
 import logging
 
+import pyignite3
 import pytest
 
-from tests.util import check_cluster_started, start_cluster_gen
+from tests.util import check_cluster_started, start_cluster_gen, server_addresses_basic
 
 logger = logging.getLogger('pyignite3')
 logger.setLevel(logging.DEBUG)
+
+
+@pytest.fixture()
+def table_name(request):
+    return request.node.originalname
+
+
+@pytest.fixture()
+def connection():
+    conn = pyignite3.connect(address=server_addresses_basic[0])
+    yield conn
+    conn.close()
+
+
+@pytest.fixture()
+def cursor(connection):
+    cursor = connection.cursor()
+    yield cursor
+    cursor.close()
 
 
 @pytest.fixture(autouse=True, scope="session")
@@ -28,3 +48,4 @@ def cluster():
         yield from start_cluster_gen()
     else:
         yield None
+
